requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
requests==2.31.0
pyyaml==6.0.1
